Rewrite Avatar
